requirements.txt
bpy
pyyaml